requirements.txt
mysql-connector-python==9.5.0
neat-python==0.92
pygame==2.6.1
python-dotenv==1.2.1
